PyDAS: Fix tests
The option '-P' to nosetests prevents it from manipulating
(and in this case messing up) the python path.

The '--all-modules -w tests' is needed instead of just giving
the path 'tests/*.py' because conda wraps that in additional
quotes '', preventing the *.py from expanding.

The '-v' just makes it verbose.

Also added a matplotlib dependency for the tests.