Fix the babel test paths so it tests the correct binaries.
This is to fix up the python scripts that test the various openbabel
executables. They were looking for the executables relative to the 
testing folder, under the assumption they were part of the usual
cmake, make, ctest, suite. But we instead want to test the binaries that 
we just installed somewhere else (eg. the anaconda _test environment).