Fixed failing build on osx.
Cmake flags are propagated to the external project builds in the
SuperBuild. The install (CMAKE_INSTALL_PREFIX) flag for ITK
caused the build to fail. This was removed as we are not doing any
install. Also turned the BUILD_EXAMPLES  flag to off (default is on),
as these are not required for building SimpleITK.