gcc-4.8: Fixes for 57068f65
"Try not using binary prefix replacement for gcc 4.8 in Linux 32"

.. is the cause of:
https://github.com/conda/conda-recipes/issues/510

This version introduces a patch to libiberty that fixes relocation
in the face of '..' and then configures various directories with
'..' so that the front part of the paths match --prefix and relocation
does not fail.

Also adds a _DEBUG variable to allow for quicker iteration, debug
symbols and lack of optimization.