gcc-4.8: add library paths reported by system linker
Right now only part of the system's libc is symlinked into the conda env - namely the startup files crtXXX.o.
In addition the path to the actual shared object libc.so also needs to be added to the spec file in case a custom
linker is used inside the conda env (e.g. because binutils is requested).